Modified class file location
- 클래스 위치 변경
