+++ app/src/main/AndroidManifest.xml
@@ -19,22 +19,22 @@
             android:theme="@style/AppTheme.NoActionBar"
             tools:ignore="GoogleAppIndexingWarning">
 
-        <activity android:name=".view.CategorySelectActivity">
+        <activity android:name=".view.activity.CategorySelectActivity">
         </activity>
-        <activity android:name=".view.MemberInfoInputActivity">
+        <activity android:name=".view.activity.MemberInfoInputActivity">
         </activity>
-        <activity android:name=".view.SignUpActivity">
+        <activity android:name=".view.activity.SignUpActivity">
         </activity>
-        <activity android:name=".view.LoginActivity">
+        <activity android:name=".view.activity.LoginActivity">
         </activity>
-        <activity android:name=".view.SplashActivity">
+        <activity android:name=".view.activity.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <activity android:name=".view.MainActivity">
+        <activity android:name=".view.activity.MainActivity">
         </activity>
     </application>
