+++ app/src/main/res/layout/activity_login.xml
@@ -8,7 +8,7 @@
         android:gravity="center_horizontal"
         android:orientation="vertical"
         android:background="@color/color_eff5f9"
-        tools:context=".view.LoginActivity">
+        tools:context=".view.activity.LoginActivity">
 
     <ImageView
             android:layout_width="132dp"
