+++ app/src/main/res/layout/activity_member_info_input.xml
@@ -7,7 +7,7 @@
         android:layout_height="match_parent"
         android:orientation="vertical"
         android:background="@color/color_eff5f9"
-        tools:context=".view.MemberInfoInputActivity">
+        tools:context=".view.activity.MemberInfoInputActivity">
     <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
             android:layout_width="match_parent"
