+++ app/src/main/res/layout/activity_sign_up.xml
@@ -7,7 +7,7 @@
         android:layout_height="match_parent"
         android:orientation="vertical"
         android:background="@color/color_eff5f9"
-        tools:context=".view.SignUpActivity">
+        tools:context=".view.activity.SignUpActivity">
 
     <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
