+++ app/src/main/res/layout/activity_splash.xml
@@ -7,7 +7,7 @@
     android:background="@drawable/splash_background"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".view.SplashActivity">
+    tools:context=".view.activity.SplashActivity">
 
     <ImageView
         android:id="@+id/iv1"
